Breakout the data file handling from stream handling
First step in allowing us to place part of this on server side.
