+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod content_sensitive_splitter;
 pub mod create;
 pub mod cuckoo_filter;
 pub mod dump_stream;
+pub mod db;
 pub mod hash;
 pub mod hash_index;
 pub mod iovec;
